# _example/auth/api/auth-openapi-spec/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Auth OpenAPI",
    "title": "Auth",
    "contact": {
      "name": "tKeel",
      "url": "https://tKeel.io/",
      "email": "auth@tkeel.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.1.0",
    "x-taggroups": []
  },
  "paths": {
...
    "/v1/tenants": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tenant"
        ],
        "summary": "get tenants",
        "operationId": "Query",
        "parameters": [
          {
            "type": "string",
            "name": "tenant_id",
            "in": "query"
          },
          {
            "type": "string",
            "name": "title",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/v1.Tenant"
              }
            }
          }
        }
      },
      "post": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tenant"
        ],
        "summary": " Create a tenant",
        "operationId": "Create",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1.TenantCreteIn"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/v1.TenantCreateOut"
            }
          }
        }
      }
    },
    "/v1/tenants/users": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tenant"
        ],
        "summary": "get users",
        "operationId": "UserQuery",
        "parameters": [
          {
            "type": "string",
            "description": "tenant`s ID",
            "name": "tenant_id",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "user`s ID",
            "name": "user_id",
            "in": "query"
          },
          {
            "type": "string",
            "description": "search by condition",
            "name": "search_key",
            "in": "query"
          },
          {
            "type": "string",
            "description": "condition`s value",
            "name": "key_words",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/v1.User"
              }
            }
          }
        }
      },
      "post": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tenant"
        ],
        "summary": "create a user",
        "operationId": "UserCreate",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1.UserCreateIn"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/v1.User"
            }
          }
        }
      }
    },
    "/v1/tenants/users/{user_id}": {
      "delete": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tenant"
        ],
        "summary": "delete a  users",
        "operationId": "UserDelete",
        "parameters": [
          {
            "type": "string",
            "name": "user_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/v1/tenants/{tenant_id}": {
      "delete": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tenant"
        ],
        "summary": "delete a tenant",
        "operationId": "Delete",
        "parameters": [
          {
            "type": "string",
            "description": "tenant's ID",
            "name": "tenant_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        }
      }
    }
  },
  "definitions": {
...
    "v1.Tenant": {
      "required": [
        "id",
        "title",
        "remark"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "remark": {
          "type": "string"
        },
        "title": {
          "type": "string"
        }
      }
    },
    "v1.TenantCreateOut": {
      "required": [
        "tenant_id",
        "title",
        "remark",
        "admin"
      ],
      "properties": {
        "admin": {
          "$ref": "#/definitions/v1.UserCreateIn"
        },
        "remark": {
          "type": "string"
        },
        "tenant_id": {
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "type": "string"
        }
      }
    },
    "v1.TenantCreteIn": {
      "required": [
        "title",
        "remark",
        "admin"
      ],
      "properties": {
        "admin": {
          "$ref": "#/definitions/v1.UserCreateIn"
        },
        "remark": {
          "type": "string"
        },
        "title": {
          "type": "string"
        }
      }
    },
...
    "v1.User": {
      "required": [
        "user_id",
        "external_id",
        "tenant_id",
        "username",
        "nick_name",
        "avatar",
        "email"
      ],
      "properties": {
        "avatar": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "external_id": {
          "type": "string"
        },
        "nick_name": {
          "type": "string"
        },
        "tenant_id": {
          "type": "integer",
          "format": "int32"
        },
        "user_id": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "v1.UserCreateIn": {
      "required": [
        "tenant_id",
        "username",
        "password",
        "nick_name",
        "avatar",
        "email"
      ],
      "properties": {
        "avatar": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "nick_name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "tenant_id": {
          "type": "integer",
          "format": "int32"
        },
        "username": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "jwt": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "jwt": []
    }
  ]
}