# apirouter/tenant/v1/register.go
/*
Copyright 2021 The tKeel Authors.
Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at
    http://www.apache.org/licenses/LICENSE-2.0
Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
*/

package v1

import (
	"net/http"

	"github.com/tkeel-io/security/apirouter"
	"github.com/tkeel-io/security/constants"
	"github.com/tkeel-io/security/errcode"

	"github.com/emicklei/go-restful"
	restfulspec "github.com/emicklei/go-restful-openapi"
)

func RegisterToRestContainer(c *restful.Container) error {
	webservice := apirouter.GetWebserviceWithPatch(c, "/v1/tenants")

	handler := newTenantHandler()

	webservice.Route(webservice.POST("").
		To(handler.Create).
		Doc(" Create a tenant").
		Reads(TenantCreteIn{}).
		Returns(http.StatusOK, errcode.ErrMsgOK, TenantCreateOut{}).
		Metadata(restfulspec.KeyOpenAPITags, []string{constants.APITagTenant}))

	webservice.Route(webservice.GET("").
		To(handler.Query).
		Doc("get tenants").
		Param(webservice.QueryParameter("tenant_id", "").Required(false)).
		Param(webservice.QueryParameter("title", "").Required(false)).
		Returns(http.StatusOK, errcode.ErrMsgOK, []Tenant{}).
		Metadata(restfulspec.KeyOpenAPITags, []string{constants.APITagTenant}))

	webservice.Route(webservice.DELETE("/{tenant_id}").
		To(handler.Delete).
		Doc("delete a tenant").
		Param(webservice.PathParameter("tenant_id", "tenant's ID").Required(true)).
		Returns(http.StatusOK, errcode.ErrMsgOK, nil).
		Metadata(restfulspec.KeyOpenAPITags, []string{constants.APITagTenant}))

	webservice.Route(webservice.POST("/users").
		To(handler.UserCreate).
		Doc("create a user").
		Reads(UserCreateIn{}).
		Returns(http.StatusOK, errcode.ErrMsgOK, User{}).
		Metadata(restfulspec.KeyOpenAPITags, []string{constants.APITagTenant}))

	webservice.Route(webservice.GET("/users").
		To(handler.UserQuery).
		Doc("get users").
		Param(webservice.QueryParameter("tenant_id", "tenant`s ID").Required(true)).
		Param(webservice.QueryParameter("user_id", "user`s ID").Required(false)).
		Param(webservice.QueryParameter("search_key", "search by condition").Required(false)).
		Param(webservice.QueryParameter("key_words", "condition`s value").Required(false)).
		Returns(http.StatusOK, errcode.ErrMsgOK, []User{}).
		Metadata(restfulspec.KeyOpenAPITags, []string{constants.APITagTenant}))

	webservice.Route(webservice.DELETE("/users/{user_id}").
		To(handler.UserDelete).
		Doc("delete a  users").
		Param(webservice.PathParameter("user_id", "").Required(true)).
		Returns(http.StatusOK, errcode.ErrMsgOK, nil).
		Metadata(restfulspec.KeyOpenAPITags, []string{constants.APITagTenant}))

	return nil
}